requirements.txt
confluent-kafka>=2.3.0
python-dotenv>=1.0.0
requests>=2.28.0
folium>=0.14.0